// src/app/components/common/user-form/user-form.interface.ts
export interface IErrorsUser {
  firstName: string
  lastName: string
  gender: string
  email: string
  password: string
  repeat_password: string
}

export interface IRegisterUser {
  firstName: string
  lastName: string
  gender: string
  email: string
  password: string
  repeat_password: string
}